+++ src/com/intellij/idea/plugin/hybris/codeInspection/rule/impex/ImpexUnknownMacrosInspection.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -20,105 +20,36 @@ package com.intellij.idea.plugin.hybris.codeInspection.rule.impex
 
 import com.intellij.codeHighlighting.HighlightDisplayLevel
 import com.intellij.codeInspection.LocalInspectionTool
-import com.intellij.codeInspection.LocalInspectionToolSession
 import com.intellij.codeInspection.ProblemHighlightType
 import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.idea.plugin.hybris.common.HybrisConstants
 import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils.message
-import com.intellij.idea.plugin.hybris.impex.psi.*
+import com.intellij.idea.plugin.hybris.impex.psi.ImpexMacroUsageDec
+import com.intellij.idea.plugin.hybris.impex.psi.ImpexVisitor
 import com.intellij.idea.plugin.hybris.impex.psi.references.ImpexMacroReference
-import com.intellij.openapi.vfs.LocalFileSystem
 import com.intellij.psi.PsiElementVisitor
-import com.intellij.psi.PsiFile
-import com.intellij.psi.PsiManager
-import com.intellij.psi.util.PsiTreeUtil
-import org.apache.commons.lang3.StringUtils
-import java.io.File
 
 class ImpexUnknownMacrosInspection : LocalInspectionTool() {
-    private val cachedMacros = HashMap<String, Boolean>()
     override fun getDefaultLevel(): HighlightDisplayLevel = HighlightDisplayLevel.ERROR
-    override fun buildVisitor(holder: ProblemsHolder, isOnTheFly: Boolean): PsiElementVisitor = UnknownMacrosVisitor(holder, cachedMacros)
-    override fun inspectionStarted(session: LocalInspectionToolSession, isOnTheFly: Boolean) {
-        cachedMacros.clear()
-        PsiTreeUtil.findChildrenOfAnyType(session.file, ImpexMacroDeclaration::class.java)
-            .map { ImpexMacroReference.escapeName(it.firstChild.text) }
-            .forEach { cachedMacros[it] = true }
-    }
+    override fun buildVisitor(holder: ProblemsHolder, isOnTheFly: Boolean): PsiElementVisitor = UnknownMacrosVisitor(holder)
 }
 
-private class UnknownMacrosVisitor(private val problemsHolder: ProblemsHolder, private val cachedMacros: HashMap<String, Boolean>) : ImpexVisitor() {
+private class UnknownMacrosVisitor(private val problemsHolder: ProblemsHolder) : ImpexVisitor() {
 
     override fun visitMacroUsageDec(usage: ImpexMacroUsageDec) {
         val text = usage.text
         if (text.startsWith(HybrisConstants.IMPEX_CONFIG_COMPLETE_PREFIX)) return
+
+        if (usage.reference?.resolve() != null) return
+
         val macroName = text
             .takeIf { it.isNotEmpty() }
             ?.let { ImpexMacroReference.escapeName(it) }
             ?: return
-
-        val isDeclarationExists = cachedMacros[macroName]
-        if (isDeclarationExists == true) return
-
-        if (isDeclarationExists != null && isDeclarationExists == false) {
-            problemsHolder.registerProblem(
-                usage,
-                message("hybris.inspections.impex.ImpexUnknownMacrosInspection.key", macroName),
-                ProblemHighlightType.ERROR
-            )
-        } else {
-            val declaration = findMacrosDeclaration(usage.containingFile, macroName)
-            if (declaration == null) {
-                cachedMacros[macroName] = false
-                problemsHolder.registerProblem(
-                    usage,
-                    message("hybris.inspections.impex.ImpexUnknownMacrosInspection.key", macroName),
-                    ProblemHighlightType.ERROR
-                )
-            } else {
-                cachedMacros[macroName] = true
-            }
-        }
+        problemsHolder.registerProblem(
+            usage,
+            message("hybris.inspections.impex.ImpexUnknownMacrosInspection.key", macroName),
+            ProblemHighlightType.ERROR
+        )
     }
-
-    private fun findMacrosDeclaration(file: PsiFile, macroName: String): ImpexMacroDeclaration? {
-        val declarations = PsiTreeUtil.findChildrenOfAnyType(file, ImpexMacroDeclaration::class.java)
-        return declarations.find { macroName.startsWith(it.firstChild.text) }
-    }
-
-    override fun visitString(o: ImpexString) {
-        super.visitString(o)
-        resolveIncludeExternalData(o)
-    }
-
-    private fun resolveIncludeExternalData(impexString: ImpexString) {
-        val text = impexString.text
-        var index = text.indexOf("impex.includeExternalData")
-        if (index == -1) return
-
-        index = text.indexOf("getResourceAsStream")
-        if (index == -1) return
-
-        val startIndex = text.indexOf("(", index)
-        if (startIndex == -1) return
-
-        val endIndex = text.indexOf(")", startIndex)
-        if (endIndex == -1) return
-
-        val resource = text.substring(startIndex + 1, endIndex)
-            .let { StringUtils.strip(it, "\"' ") }
-        val directory = impexString.containingFile.containingDirectory.virtualFile
-        val dirPath = directory.canonicalPath
-        val referencedFile = LocalFileSystem.getInstance().findFileByIoFile(File(dirPath, resource))
-
-        if (referencedFile == null || !referencedFile.exists()) return
-
-        val referencedPsi = PsiManager.getInstance(impexString.project).findFile(referencedFile) as? ImpexFile ?: return
-        val declarations = PsiTreeUtil.findChildrenOfAnyType(referencedPsi, ImpexMacroDeclaration::class.java)
-        declarations.forEach { declaration -> cachedMacros[declaration.firstChild.text] = true }
-        val impString = PsiTreeUtil.findChildrenOfAnyType(referencedPsi, ImpexString::class.java)
-        impString.forEach { resolveIncludeExternalData(it) }
-    }
-
-
 }